requirements.txt
einops==0.8.1
diffusers==0.35.1
huggingface-hub==0.34.4
hydra-core==1.3.2
pytorch-lightning==2.5.3
apex==0.9.10.dev0
deepspeed==0.17.5
lightning==2.5.3
pytorchvideo==0.1.5
timm==1.0.19
transformers==4.55.0
wandb==0.21.1
colorama==0.4.6
scipy==1.16.1
wandb_osh==1.2.2
mkl==2024.0
wandb[media]